Implement DFS pathfinding algorithm, integrate pathfinding algorithm into application
This commit introduces a Depth-First Search (DFS) pathfinding function. The function is designed to traverse a graph structure to find a path from a start node to an end node using the Depth-First Search algorithm.
The other change is the integration of the pathfinding
algorithms to the main application, when the visualze
button is clicked. There is a function called runPathfinding
that will dynamically import the module by the name of the algorithm,
algorithm names are defined in constants.ts. Then the algorithm will be
accessed by using the 'default' property of the module. Then the
function will be called with the city, startNode, and endNode passed as
parameters. It's important to await this function because the pathfinding
functions asynchronously load the data. See how that is done in the dfs
module.

Changes included:
- Added a new file containing the DFS pathfinding function.
- Implemented DFS traversal logic.
- Implemented proper path reconstruction
- Implemented function that handles pathfinding after visualize button is clicked

issue #15