Add distance and maxSpeed information to adjacent nodes in adjacency list data.
This was done by updating the map_parser.py script
to calculate the Haversine distance, and extract
the maxSpeed information from Way tags. This information
is now incorporated in our data json files. The maxSpeed
may be null, as some Way elements do not have a maxSpeed
defined. Now each node has the informatin for the distance
of the road to the adjacent node as well as the maxSpeed
of that road.

I also updated the nodeInfo type in types.ts to reflect
this change in our data, and created a new type called
adjacencyInformation that encapsulate the information
of adjacent nodes.

issue #18