Modify map_parser.py script to properly handle one-way streets. Implement directional parsing logic to ensure accurate representation of street networks.
We check if the Way is tagged as a one_way or not.
For every two adjacent nodes in the list of nodes
of a Way, node1 and node2. If the Way is a one_way
There exists the directed edge (node1, node2), if
it is not a one way, there exists the directed
edges (node1, node2) and (node2, node1). We add the
corresponding adjacent nodes to the graph based
on this property.

issue #18