Make pathfinding asynchronous using WebWorker
I made the pathfinding asynchronous because
pathfinding can take a long time, and since Node.json
is single-threaded it would block the app from
doing other things until the pathfinding is complete.
I'm using a WebWorker, when you click Visualize,
in the click handler function runPathfinding, it
sends a message to the Worker with the information
needed to execute the pathfinding. Note the Worker
thread does not have access to memory or objects in
the main thread, and so loads the data for itself.

issue #28