Adjust algorithms to account for no path
This fix addresses an issue in the minIndexHeap data structure, to do a bounds check on the number of children that a node has, as well as prevent a single node from being added when there is no valid path.

Due to the nature of our map having bounds, there are some nodes that are unreachable, for example if you are going east on a highway and the map suddenly cuts off, unless there is an exit in the bounded area, you will be unable to access the rest of the nodes.